backend/requirements.txt
supabase==2.15.2
stripe==11.4.1
pytest==8.3.5
pytest-mock==3.14.0
pytest-asyncio==0.26.0
annotated-doc==0.0.4
newspaper3k==0.2.8
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
click==8.1.8
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
lxml==6.0.4
trafilatura
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.1
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
